requirements.txt
Flask
Flask-Cors
jsonify
tiktoken
openai
orjson
numpy
scikit-learn
setuptools
tqdm
flask-socketio
colorlog
pillow